requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
altair>=5.0.0
plotly>=5.0.0
python-dotenv>=1.0.0
pyjwt>=2.0.0
cryptography>=41.0.0
pymupdf>=1.22.0
docx2txt>=0.8
pytesseract>=0.3.10
pillow>=10.0.0
textstat>=0.7.0
vaderSentiment>=3.3.0
nltk>=3.8.0
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
faiss-cpu>=1.7.0
huggingface-hub>=0.16.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
streamlit-autorefresh>=0.1.0
numpy>=1.24.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1